+++ src/main/java/de/bitwars/api/resources/GamesResource.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2024 Andreas Heinrich
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package de.bitwars.api.resources;
 
 import de.bitwars.api.interfaces.GamesApi;
@@ -20,12 +36,12 @@ public class GamesResource implements GamesApi {
     }
 
     @Override
-    public void deleteGame(Integer gameId) {
+    public void deleteGame(long gameId) {
 
     }
 
     @Override
-    public Game getGameById(Integer gameId) {
+    public Game getGameById(long gameId) {
         return null;
     }
 
@@ -35,27 +51,27 @@ public class GamesResource implements GamesApi {
     }
 
     @Override
-    public List<Player> listPlayersInGame(Integer gameId) {
+    public List<Player> listPlayersInGame(long gameId) {
         return List.of();
     }
 
     @Override
-    public void removePlayerFromGame(Integer gameId, Integer playerId) {
+    public void removePlayerFromGame(long gameId, long playerId) {
 
     }
 
     @Override
-    public Game startGame(Integer gameId, GameOptions gameOptions) {
+    public Game startGame(long gameId, GameOptions gameOptions) {
         return null;
     }
 
     @Override
-    public Game stopGame(Integer gameId) {
+    public Game stopGame(long gameId) {
         return null;
     }
 
     @Override
-    public Game updateGame(Integer gameId, Game game) {
+    public Game updateGame(long gameId, Game game) {
         return null;
     }
 }
